refactor(sdk): split subprocess collector
